Add support for multiple nodes
Swift can now be deployed on multiple nodes. There are
two new containers: ring-master and a ring-minion.
The ring master creates the rings and makes it available over HTTP.
The ring-minion are deployed on the same pods as the proxy and the
storage containers. The minions download the rings from the master
and stores where the swift processes expects them.